chore: add import linter and do some modules arrangements
