+++ appointment_confirmation_module/internal/core/input_ports/appointment_notification_port.py
@@ -1,6 +1,6 @@
 from abc import abstractmethod, ABC
 
-from appointment_confirmation_module.core.models.appointment_notification import AppointmentNotification
+from appointment_confirmation_module.internal.core.models.appointment_notification import AppointmentNotification
 
 
 class AppointmentNotificationPort(ABC):
