+++ appointment_confirmation_module/internal/core/models/appointment_notification.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
-from appointment_confirmation_module.core.models.notification import Notification
-from appointment_confirmation_module.core.models.slot import Slot
+from appointment_confirmation_module.internal.core.models.notification import Notification
+from appointment_confirmation_module.internal.core.models.slot import Slot
 
 
 @dataclass
