+++ appointment_confirmation_module/internal/core/models/slot.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
 from datetime import datetime
 
-from appointment_confirmation_module.core.models.doctor import Doctor
-from appointment_confirmation_module.core.models.patient import Patient
+from appointment_confirmation_module.internal.core.models.doctor import Doctor
+from appointment_confirmation_module.internal.core.models.patient import Patient
 
 
 @dataclass
