+++ appointment_confirmation_module/internal/core/output_ports/appointment_notification_repo_abc.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod, ABC
+
+from appointment_confirmation_module.internal.core.models.appointment_notification import AppointmentNotification
+from appointment_confirmation_module.internal.core.models.notification import Notification
+
+
+class NotificationRepoAbc(ABC):
+    @abstractmethod
+    def save(self, notification: Notification) -> None:
+        pass
