+++ appointment_confirmation_module/internal/core/output_ports/notification_providers/notification_provider_port.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from appointment_confirmation_module.core.models.notification import Notification
+from appointment_confirmation_module.internal.core.models.notification import Notification
 
 
 class NotificationProviderPort(ABC):
