+++ appointment_confirmation_module/internal/shell/views.py
