+++ appointment_confirmation_module/shared/__init__.py
@@ -1,0 +1,4 @@
+from .api import send_appointment_notification
+from .doctor_dto import DoctorDTO
+from .patient_dto import PatientDTO
+from .slot_dto import SlotDTO
