+++ src/settings.py
@@ -40,7 +40,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'django_extensions',
     'doctor_availability.application.apps.DoctorAvailabilityConfig',
-    'appointment_confirmation_module.shell.apps.AppointmentConfirmationModuleConfig',
+    'appointment_confirmation_module.internal.shell.apps.AppointmentConfirmationModuleConfig',
     'patients_appointments.apps.PatientsAppointmentsConfig',
     'doctor_appointment_management.apps.DoctorAppointmentManagementConfig',
     'shared.apps.SharedConfig',
@@ -131,7 +131,7 @@ DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
 MIGRATION_MODULES = {
     'doctor_availability': 'doctor_availability.infrastructure.migrations',
-    'appointment_confirmation_module': 'appointment_confirmation_module.shell.db.migrations',
+    'appointment_confirmation_module': 'appointment_confirmation_module.internal.shell.db.migrations',
     "patients_appointments": "patients_appointments.infrastructure.db.migrations",
     "doctor_appointment_management": "doctor_appointment_management.migrations",
 }
